+++ src/components/dialog/index.ts
@@ -1,4 +1,5 @@
 export { AppDialog } from "./AppDialog";
+export { BottomSheet } from "./BottomSheet";
 export { DialogActions } from "./DialogActions";
 export type {
   AppDialogProps,
